--- a/db/flush_scheduler.cc
+++ b/db/flush_scheduler.cc
@@ -11,6 +11,11 @@
 
 namespace rocksdb {
 
+FlushScheduler::FlushScheduler()
+{
+  is_empty_ = true;
+} 
+
 void FlushScheduler::ScheduleFlush(ColumnFamilyData* cfd) {
 #ifndef NDEBUG
   assert(column_families_set_.find(cfd) == column_families_set_.end());
@@ -18,6 +23,7 @@ void FlushScheduler::ScheduleFlush(ColumnFamilyData* cfd) {
 #endif  // NDEBUG
   cfd->Ref();
   column_families_.push_back(cfd);
+  is_empty_ = false;
 }
 
 ColumnFamilyData* FlushScheduler::GetNextColumnFamily() {
@@ -34,6 +40,7 @@ ColumnFamilyData* FlushScheduler::GetNextColumnFamily() {
       break;
     }
   }
+  is_empty_ = column_families_.empty();  
 #ifndef NDEBUG
   if (cfd != nullptr) {
     auto itr = column_families_set_.find(cfd);
@@ -44,7 +51,7 @@ ColumnFamilyData* FlushScheduler::GetNextColumnFamily() {
   return cfd;
 }
 
-bool FlushScheduler::Empty() { return column_families_.empty(); }
+bool FlushScheduler::Empty() { return is_empty_.load(); }
 
 void FlushScheduler::Clear() {
   for (auto cfd : column_families_) {
@@ -58,6 +65,7 @@ void FlushScheduler::Clear() {
     }
   }
   column_families_.clear();
+  is_empty_ = true;
 }
 
 }  // namespace rocksdb