--- a/db/memtable_allocator.h
+++ b/db/memtable_allocator.h
@@ -14,14 +14,13 @@
 #include "util/allocator.h"
 
 namespace rocksdb {
-
-class Arena;
+ 
 class Logger;
 class WriteBuffer;
 
 class MemTableAllocator : public Allocator {
  public:
-  explicit MemTableAllocator(Arena* arena, WriteBuffer* write_buffer);
+  explicit MemTableAllocator(Allocator* allocator, WriteBuffer* write_buffer);
   ~MemTableAllocator();
 
   // Allocator interface
@@ -35,7 +34,7 @@ class MemTableAllocator : public Allocator {
   void DoneAllocating();
 
  private:
-  Arena* arena_;
+  Allocator* allocator_;
   WriteBuffer* write_buffer_;
   size_t bytes_allocated_;