The goal of this patch is to improve write performance by permitting concurrent execution of write operations. It includes concurrency control that synchronizes between concurrent write operations, compactions, and snapshot creations.
This patch is a simple variant of the algorithm presented in the EuroSys'2015 paper: "Scaling Concurrent Log-Structured Data Stores" . It supports concurrent write operations + WAL + consistent snapshots.
The patch is enabled when DBOptions::allow_concurrent_write_operations== true

Here are some performance results on a Xeon E5620 machine with 2 quad-core CPUs:

Before:
-------
./db_bench  --benchmarks=fillrandom --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=16000000 --threads=1
fillrandom   :       4.571 micros/op 218747 ops/sec;   24.2 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=8000000 --threads=2
fillrandom   :       9.831 micros/op 101717 ops/sec;   11.3 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=4000000 --threads=4
fillrandom   :      10.660 micros/op 93812 ops/sec;   10.4 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=2000000 --threads=8
fillrandom   :      11.087 micros/op 90196 ops/sec;   10.0 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=1000000 --threads=16
fillrandom   :      11.328 micros/op 88276 ops/sec;    9.8 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=16000000 --threads=1
fillrandom   :       7.541 micros/op 132602 ops/sec;   14.7 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=8000000 --threads=2
fillrandom   :      14.022 micros/op 71315 ops/sec;    7.9 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=4000000 --threads=4
fillrandom   :      13.898 micros/op 71952 ops/sec;    8.0 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=2000000 --threads=8
fillrandom   :      13.561 micros/op 73741 ops/sec;    8.2 MB/s

./db_bench  --benchmarks=fillrandom --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=1000000 --threads=16
fillrandom   :      13.049 micros/op 76635 ops/sec;    8.5 MB/s


After:
-------
./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=16000000 --threads=1
fillrandom   :       3.624 micros/op 275972 ops/sec;   30.5 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=8000000 --threads=2
fillrandom   :       2.444 micros/op 409086 ops/sec;   45.3 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=4000000 --threads=4
fillrandom   :       1.499 micros/op 667317 ops/sec;   73.8 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=2000000 --threads=8
fillrandom   :       1.235 micros/op 809447 ops/sec;   89.5 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=1 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=1000000 --threads=16
fillrandom   :       1.234 micros/op 810624 ops/sec;   89.7 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=16000000 --threads=1
fillrandom   :       4.249 micros/op 235328 ops/sec;   26.0 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=8000000 --threads=2
fillrandom   :       3.263 micros/op 306432 ops/sec;   33.9 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=4000000 --threads=4
fillrandom   :       2.908 micros/op 343906 ops/sec;   38.0 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=2000000 --threads=8
fillrandom   :       2.623 micros/op 381304 ops/sec;   42.2 MB/s

./db_bench  --benchmarks=fillrandom --memtablerep=lock_free_skip_list --concurrent_writes=1 --disable_wal=0 --write_buffer_size=134217728 --max_background_compactions=10 --max_write_buffer_number=15 --mmap_write=1 --num=1000000 --threads=16
fillrandom   :       2.995 micros/op 333911 ops/sec;   36.9 MB/s