@siying 2015-05-21
Since it is only for a feature branch "write_throughput", I'll just merge it and we can improve from there.