@siying 2015-05-22
Oops, I just realized that the description is only pasted in comments but not in your commit message.